notebook/requirements_section-2-serverless.txt
-e .
aws-cdk.core>=1.3.0
aws-cdk.aws_ec2

aws-cdk.aws-autoscaling
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_ecs
aws-cdk.aws_ecs_patterns
aws-cdk.aws_eks
aws-cdk.aws_elasticloadbalancingv2
aws_cdk.aws_rds
